Right pointer of string view of a node is removed partially.
In the end of building of a suffix tree a pointer of a node which points
to the end of the substring of the node must point to the end of entire
string. From technical point there is no reason to update it implicitly
as the building goes on. `implicitly` means keeping a weak pointer to
the current position the input string is processed before.