Position enum is removed.
There is no need to maintain the type as a node stores just a value
instead of pointer. If a node carries suffix to the current known end of
input string it points to its end. Such node is a leaf node. There is no
reason from technical point of view to update current end in leaf nodes.
Splitting a node is the only case when the right pointer is involved. In
the case it's copied to one of the nodes a splitted one. So if the
splitted node was a leaf the new one will be a leaf too so it's
reasonable to put just the length of the input string.